pyproject.toml
[project]
name = "system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cairocffi==1.7.1",
    "cairosvg==2.7.1",
    "cffi==1.17.1",
    "cssselect2==0.7.0",
    "defusedxml==0.7.1",
    "mouseinfo==0.1.3",
    "numpy==2.1.3",
    "pandas==2.2.3",
    "pillow==11.0.0",
    "psutil==6.1.0",
    "pyautogui==0.9.54",
    "pycparser==2.22",
    "pygetwindow==0.0.9",
    "pymsgbox==1.0.9",
    "pyperclip==1.9.0",
    "pyrect==0.2.0",
    "pyscreeze==1.0.1",
    "pyserial==3.5",
    "pystray==0.19.5",
    "python-dateutil==2.9.0.post0",
    "pytweening==1.2.0",
    "pytz==2024.2",
    "pywin32==308",
    "pyyaml==6.0.2",
    "six==1.16.0",
    "tinycss2==1.4.0",
    "tzdata==2024.2",
    "webencodings==0.5.1",
]
